requirements.txt
numpy
torch
torchmetrics
torchvision
scipy
Pillow
opencv-python
plyfile
tqdm